Add rocm_kernel.h, updates.
