+++ onnxruntime/core/providers/rocm/rocm_fwd.h
@@ -3,6 +3,8 @@
 
 #pragma once
 
+#include "core/framework/op_kernel.h"
+
 namespace onnxruntime {
 namespace rocm {
 template <typename T>
